+++ recommender/api/schemas/recommendation.py
@@ -100,7 +100,7 @@ recommendation_context = api.model(
             description="Version of the recommendation engine",
             example="NCF",
         ),
-        "elastic_services": fields.List(
+        "candidates": fields.List(
             fields.Integer(
                 title="Service ID", description="The unique identifier of the service"
             ),
@@ -116,6 +116,12 @@ recommendation_context = api.model(
 recommendation = api.model(
     "Recommendations",
     {
+        "panel_id": fields.String(
+            required=True,
+            title="Root type",
+            description="The unique identifier of the recommender panel on the page",
+            example="v1",
+        ),
         "recommendations": fields.List(
             fields.Integer(
                 title="Service ID", description="The unique identifier of the service"
@@ -124,6 +130,51 @@ recommendation = api.model(
             title="Recommended services list",
             description="List of the recommended services' IDs",
             example=[1234, 2345, 3456],
-        )
+        ),
+        "explanations": fields.List(
+            fields.String(
+                title="Explanation",
+                description="Explanation of choice of the corresponding service",
+            ),
+            required=True,
+            title="Explanations list",
+            description="List of the recommended services explanations",
+            example=[
+                "some long explanation",
+                "some long explanation",
+                "some long explanation",
+            ],
+        ),
+        "explanations_short": fields.List(
+            fields.String(
+                title="Short explanation",
+                description="Short explanation of choice of the corresponding service",
+            ),
+            required=True,
+            title="Short explanations list",
+            description="List of the recommended services short explanations",
+            example=[
+                "some short explanation",
+                "some short explanation",
+                "some short explanation",
+            ],
+        ),
+        "scores": fields.List(
+            fields.Float(
+                title="Service score",
+                description="Score of the corresponding service on the basis"
+                " of which the choice of recommendation has been made",
+            ),
+            required=True,
+            title="Recommended services scores list",
+            description="List of the recommended services scores",
+            example=[0.7, 0.2, 0.1],
+        ),
+        "engine_version": fields.String(
+            required=True,
+            title="Engine version",
+            description="Version of the recommendation engine",
+            example="NCF",
+        ),
     },
 )