+++ recommender/engines/ncf/inference/ncf_inference_component.py
@@ -8,6 +8,7 @@ from typing import List, Tuple
 import torch
 
 from recommender.engines.base.base_inference_component import MLEngineInferenceComponent
+from recommender.engines.explanations import Explanation
 from recommender.engines.ncf.ml_components.neural_collaborative_filtering import (
     NeuralCollaborativeFilteringModel,
     NEURAL_CF,
@@ -25,10 +26,16 @@ logger = get_logger(__name__)
 
 class NCFInferenceComponent(MLEngineInferenceComponent):
     """
-    Recommender engine that provides logged-in users with personalized recommendations in a given context.
+    Recommender engine that provides logged-in users with personalized
+     recommendations in a given context.
     """
 
     engine_name = "NCF"
+    default_explanation = Explanation(
+        long="This service has been selected based on historical orders of the"
+        " given user and similar orders of similar users",
+        short="This service has been selected based on historical orders.",
+    )
 
     def __init__(self, K: int) -> None:
         self.neural_cf_model = None
@@ -78,13 +85,13 @@ class NCFInferenceComponent(MLEngineInferenceComponent):
         return users_ids, users_tensor, services_ids, services_tensor
 
     def _get_ranking(
-        self, user: User, elastic_services: Tuple[int], search_data: SearchData
-    ) -> List[Tuple[float, int]]:
+        self, user: User, candidates: Tuple[int], search_data: SearchData
+    ) -> Tuple[List[int], List[float]]:
         """Generate services ranking.
 
         Args:
             user: user for whom recommendation will be generated.
-            elastic_services: item space from the Marketplace.
+            candidates: item space from the Marketplace.
             search_data: search phrase and filters information for narrowing
              down an item space.
 
@@ -93,9 +100,7 @@ class NCFInferenceComponent(MLEngineInferenceComponent):
         """
 
         candidate_services = list(
-            retrieve_services_for_recommendation(
-                elastic_services, user.accessed_services
-            )
+            retrieve_services_for_recommendation(candidates, user.accessed_services)
         )
         if len(candidate_services) < self.K:
             raise InsufficientRecommendationSpaceError()
@@ -116,24 +121,30 @@ class NCFInferenceComponent(MLEngineInferenceComponent):
             list(zip(matching_probs, candidate_services_ids)), reverse=True
         )
 
-        return ranking
+        scores, recommended_services_ids = [list(t) for t in list(zip(*ranking))]
+
+        return recommended_services_ids, scores
 
     def _generate_recommendations(
-        self, user: User, elastic_services: Tuple[int], search_data: SearchData
-    ) -> List[int]:
+        self, user: User, candidates: Tuple[int], search_data: SearchData
+    ) -> Tuple[List[int], List[float], List[Explanation]]:
         """Generate recommendation for logged user.
 
         Args:
             user: user for whom recommendation will be generated.
-            elastic_services: item space from the Marketplace.
+            candidates: item space from the Marketplace.
             search_data: search phrase and filters information for narrowing
              down an item space.
 
         Returns:
             recommended_services_ids: List of recommended services ids.
+            scores: List of ranking scores for all recommended services.
+            explanations: List of explanations for all recommended services.
         """
 
-        top_k = self._get_ranking(user, elastic_services, search_data)[: self.K]
-        recommended_services_ids = [pair[1] for pair in top_k]
+        recommended_services_ids, scores = tuple(
+            l[: self.K] for l in self._get_ranking(user, candidates, search_data)
+        )
+        explanations = self._generate_explanations()
 
-        return recommended_services_ids
+        return recommended_services_ids, scores, explanations