+++ recommender/engines/rl/inference/rl_inference_component.py
@@ -1,7 +1,7 @@
 # pylint: disable=missing-module-docstring, missing-class-docstring, too-few-public-methods
 # pylint: disable=too-many-instance-attributes, line-too-long
 
-from typing import Tuple
+from typing import Tuple, List
 
 import torch
 
@@ -10,6 +10,7 @@ from recommender.engines.autoencoders.ml_components.embedder import (
     USER_EMBEDDER,
     SERVICE_EMBEDDER,
 )
+from recommender.engines.explanations import Explanation
 from recommender.engines.panel_id_to_services_number_mapping import K_TO_PANEL_ID
 from recommender.engines.rl.utils import create_state
 from recommender.engines.base.base_inference_component import MLEngineInferenceComponent
@@ -21,10 +22,16 @@ from recommender.models import User, SearchData
 
 class RLInferenceComponent(MLEngineInferenceComponent):
     """
-    Recommender engine that provides logged-in users with personalized recommendations in a given context.
+    Recommender engine that provides logged-in users with personalized
+     recommendations in a given context.
     """
 
     engine_name = "RL"
+    default_explanation = Explanation(
+        long="This service has been selected based on the analysis of"
+        " historical services orders and users activity patterns.",
+        short="This service has been selected based on the users activity.",
+    )
 
     def __init__(self, K: int, exploration: bool = False, act_noise: float = 0.0):
         self.exploration = exploration
@@ -48,13 +55,13 @@ class RLInferenceComponent(MLEngineInferenceComponent):
         self.service_selector = ServiceSelector(self.service_embedder)
 
     def _generate_recommendations(
-        self, user: User, elastic_services: Tuple[int], search_data: SearchData
-    ) -> Tuple[int]:
+        self, user: User, candidates: Tuple[int], search_data: SearchData
+    ) -> Tuple[List[int], List[float], List[Explanation]]:
         """Generate recommendation for logged user.
 
         Args:
             user: user for whom recommendation will be generated.
-            elastic_services: item space from the Marketplace.
+            candidates: item space from the Marketplace.
             search_data: search phrase and filters information for narrowing
              down an item space.
 
@@ -62,12 +69,13 @@ class RLInferenceComponent(MLEngineInferenceComponent):
             service_ids: List of recommended services ids.
         """
 
-        state = create_state(user, elastic_services, search_data)
+        state = create_state(user, candidates, search_data)
         state_tensors = self.state_encoder([state])
         weights_tensor = self._get_weights(state_tensors)
         services_mask = self._get_service_mask(state_tensors)
-        service_ids = self.service_selector(weights_tensor, services_mask)
-        return service_ids
+        service_ids, scores = self.service_selector(weights_tensor, services_mask)
+        explanations = self._generate_explanations()
+        return service_ids, scores, explanations
 
     @staticmethod
     def _get_service_mask(state_tensors):