+++ recommender/services/deserializer.py
@@ -49,7 +49,7 @@ class Deserializer:
             panel_id=json_dict.get("panel_id"),
             engine_version=json_dict.get("engine_version"),
             services=json_dict.get("services"),
-            elastic_services=json_dict.get("elastic_services"),
+            candidates=json_dict.get("candidates"),
             search_data=search_data,
         )