+++ recommender/services/provide_service_ctx.py
@@ -15,18 +15,18 @@ def get_all_collection_ids(collection: Type[MarketplaceDocument]) -> List[int]:
 
 
 def service_ctx(body_request: Dict) -> Dict:
-    """Provide service context as "all services" if elastic_services are not specified
+    """Provide service context as "all services" if candidates are not specified
     in the recommendations request's body, for the purpose of the User Dashboard"""
 
-    elastic_services = body_request.get("elastic_services")
+    candidates = body_request.get("candidates")
     page_id = body_request.get("page_id")
 
-    if elastic_services is None:
+    if candidates is None:
         if page_id == "/dashboard":  # case for the UD
-            body_request["elastic_services"] = get_all_collection_ids(Service)
+            body_request["candidates"] = get_all_collection_ids(Service)
         else:
             logger.error(
-                "Elastic_services not provided. Only for the context of page_id == '/dashboard' they are NOT required"
+                "Candidates not provided. Only for the context of page_id == '/dashboard' they are NOT required"
             )
             raise ServicesContextNotProvidedError()