+++ tests/endpoints/conftest.py
@@ -105,7 +105,7 @@ def recommendation_data():
         "page_id": "some_page_identifier",
         "panel_id": "v1",
         "engine_version": "NCF",
-        "elastic_services": [1, 2, 3],
+        "candidates": [1, 2, 3],
         "search_data": {
             "q": "Cloud GPU",
             "categories": [1],
@@ -132,7 +132,7 @@ def recommendation_data_with_aai_uid():
         "page_id": "some_page_identifier",
         "panel_id": "v1",
         "engine_version": "NCF",
-        "elastic_services": [1, 2, 3],
+        "candidates": [1, 2, 3],
         "search_data": {
             "q": "Cloud GPU",
             "categories": [1],
@@ -158,7 +158,7 @@ def recommendation_data_with_aai_uid():
         "page_id": "some_page_identifier",
         "panel_id": "v1",
         "engine_version": "NCF",
-        "elastic_services": [1, 2, 3],
+        "candidates": [1, 2, 3],
         "search_data": {
             "q": "Cloud GPU",
             "categories": [1],