+++ tests/endpoints/test_recommendations.py
@@ -3,6 +3,8 @@
 import json
 from copy import deepcopy
 
+from recommender.engines.explanations import Explanation
+
 
 def test_recommendations(
     client, mocker, recommendation_data, recommendation_data_with_aai_uid
@@ -18,8 +20,14 @@ def test_recommendations(
         deserializer_mock = mocker.patch(
             "recommender.services.deserializer.Deserializer.deserialize_recommendation"
         )
-        mocked_recommended_services = [1, 2, 3]
-        inference_component_call_mock.return_value = mocked_recommended_services
+        mock_recommended_services_ids = [1, 2, 3]
+        mock_scores = [0.7, 0.2, 0.1]
+        mock_explanations = 3 * [Explanation(long="mock_long", short="mock_short")]
+        inference_component_call_mock.return_value = (
+            mock_recommended_services_ids,
+            mock_scores,
+            mock_explanations,
+        )
 
         response = client.post(
             "/recommendations",
@@ -28,8 +36,20 @@ def test_recommendations(
         )
 
         deserializer_data = deepcopy(recc_data)
-        deserializer_data["services"] = mocked_recommended_services
+        deserializer_data["services"] = mock_recommended_services_ids
 
         inference_component_call_mock.assert_called_once_with(recc_data)
         deserializer_mock.assert_called_once_with(deserializer_data)
-        assert response.get_json() == {"recommendations": mocked_recommended_services}
+
+        mock_explanations_long, mock_explanations_short = [
+            list(t) for t in list(zip(*[(e.long, e.short) for e in mock_explanations]))
+        ]
+
+        assert response.get_json() == {
+            "panel_id": recc_data["panel_id"],
+            "recommendations": mock_recommended_services_ids,
+            "explanations": mock_explanations_long,
+            "explanations_short": mock_explanations_short,
+            "scores": mock_scores,
+            "engine_version": recc_data["engine_version"],
+        }