+++ tests/engines/rl/inference/test_rl_inference_component.py
@@ -21,31 +21,31 @@ def test_known_user(
     assert isinstance(inference_component.engine_name, str)
 
     user = random.choice(list(User.objects))
-    elastic_services = [service.id for service in Service.objects]
+    candidates = [service.id for service in Service.objects]
 
     context = {
         "user_id": {
             "panel_id": ver,
-            "elastic_services": elastic_services,
+            "candidates": candidates,
             "search_data": {},
             "user_id": user.id,
         },
         "aai_uid": {
             "panel_id": ver,
-            "elastic_services": elastic_services,
+            "candidates": candidates,
             "search_data": {},
             "aai_uid": user.aai_uid,
         },
     }[user_id_type]
 
-    services_ids_1 = inference_component(context)
+    services_ids_1, _, _ = inference_component(context)
 
     assert isinstance(services_ids_1, list)
     assert len(services_ids_1) == PANEL_ID_TO_K.get(context["panel_id"])
     assert all([isinstance(service_id, int) for service_id in services_ids_1])
-    assert all(service in elastic_services for service in services_ids_1)
+    assert all(service in candidates for service in services_ids_1)
 
-    services_ids_2 = inference_component(context)
+    services_ids_2, _, _ = inference_component(context)
     assert services_ids_1 == services_ids_2
 
 
@@ -53,11 +53,11 @@ def test_user_cannot_be_identified(
     mongo, generate_users_and_services, mock_rl_pipeline_exec
 ):
     rl_inference_component = RLInferenceComponent(3)
-    elastic_services = [service.id for service in Service.objects]
+    candidates = [service.id for service in Service.objects]
 
     context = {
         "panel_id": "v1",
-        "elastic_services": elastic_services,
+        "candidates": candidates,
         "search_data": {},
     }